backend/requirements.txt
fastapi
uvicorn
dnspython
pydantic
python-dotenv
requests
PySocks
